requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
pyTelegramBotAPI==4.8.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
telebot==0.0.4
telebotapi==0.9
telegram==0.0.1
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
yarl==1.8.1
schedule~=1.1.0
Pillow~=9.3.0